+++ CHANGELOG.rst
@@ -4,6 +4,8 @@
 Changelog
 =========
 
+* Allow to pass the command parser as an argument.
+
 * :release:`0.1.1 <2019-10-03>`
 * Fix tag version.